# Make abbr.fish nicer and faster