--- a/share/functions/abbr.fish
+++ b/share/functions/abbr.fish
@@ -141,8 +141,7 @@ function __fish_abbr_get_by_key
 	end
 	set -l count (count $fish_user_abbreviations)
 	if test $count -gt 0
-		set -l key
-		__fish_abbr_parse_entry $argv[1] key
+		set -l key $argv[1] # This assumes the key is valid and pre-parsed
 		set -l IFS \n # ensure newline splitting is enabled
 		for i in (seq $count)
 			set -l key_i