As we've seen in #1976, users often have a lot of aliases from previous shells, and replacing them with abbreviations is sometimes too slow, especially on startup.

This PR attempts to improve the situation. With it, sourcing https://github.com/simnalamburt/pkg-cgitc/blob/28cb2b905d2c389464d79e58f0cdbfb2d78c74f1/init.fish (s/alias/abbr) goes from about 4 seconds to about 2 seconds with a cleared $fish_user_abbreviations. Still not quite snappy but better than before.

The case with pre-existing abbreviations isn't improved as much (750ms -> 500ms), but it still helps a bit.

In addition to that, there's a few cases in abbr where `string` comes in handy, replacing weird IFS tricks and handrolled escaping.

While this doesn't do any crazy stuff (like caching the keys separately for quicker checking, which could work but would be tricky to pull off if users are allowed to fiddle with $fish_user_abbreviations), it's still possible this changes behavior in edgecases. I've already checked one particular suspect - abbrs with newlines in the value.