@faho 2016-01-04
This should now work (once github notices the new commits). Sorry, I forgot part of a commit.

There's one behavioral change though: abbreviations starting with "--" are no longer quoted in the `abbr --show` output, but since they are still prefixed with "--" (i.e. the output looks like `abbr -- --__abbr3 xyz` instead of `abbr -- '--__abbr3' xyz`) there's still no ambiguity.