@krader1961 2016-01-05
LGTM although I wish there was more use of single rather double-quoted strings where the latter isn't needed. A few more milliseconds might be saved by using single-quotes where it is known that variable interpolation should not be done. But that battle is akin to a Don Quixote windmill :-)