@pickfire 2016-01-05
@faho, what about setting a lot of universal variables, will it be slowed?