@faho 2016-01-05
> I wish there was more use of single rather double-quoted strings where the latter isn't needed. A few more milliseconds might be saved by using single-quotes where it is known that variable interpolation should not be done.

@krader1961: I don't think you'll actually get any measurable change from that.

>what about setting a lot of universal variables, will it be slowed?

@pickfire: I'm not sure what you mean. The abbr system uses one universal variable - $fish_user_abbreviations (and it also allows it to be defined as global instead of universal). It doesn't touch any others, and setting this one will be just as fast with this change as without it.