@pickfire 2016-01-06
@faho: I mean when you set a lot of variable with `set -Ux`, will fish be slowed?