@ridiculousfish 2016-01-07
@faho, this looks fine to me; feel free to merge as you see fit. Thanks!