@faho 2016-01-07
Good thing I've been replaying Diablo 2 - I've actually found a tiny bug, and a slightly bigger pre-existing one:

```fish
abbr -a d2 env WINEPREFIX=/home/alfa/.wine32/ wine ~/.wine/drive_c/Program\ Files\ \(x86\)/Diablo\ II/Diablo\ II.exe
```

will 

- Read the key as "d2 env WINEPREFIX" because it's detecting the "=" as separator - this needs regex-matching instead of globs (since we need to see if there's a space before a "=", so we need a way to say "not-space") with this, but not without - this causes `abbr --show` output to show `abbr 'd2 env WINEPREFIX' /home/alfa...`

- Save the "(x86)" with unescaped parens, meaning simply handing the path like completion hands it to us leads to an unexecutable alias (compare #952) - it either needs to be double-escaped (which is annoying) or to be fixed in abbr (with some `string escape`)

I'll fix the former now since it's caused by this, and I'll check on the latter later.