@chemzqm 2016-01-14
I took me sometime to figure out the abbrs slow down the fish a lot, do hope this PR get merged.