@faho 2016-01-14
This has now been merged as ac54594..ed1919b.