@faho 2016-01-14
@chemzqm: This is very much a bandaid - the speed difference is a factor of two for a large number of abbr being added and about 1/3 for a large number of abbr being checked (i.e. already added).

A bigger speed difference is for you to not add them on any startup since they are stored in a universal variable.