@chemzqm 2016-01-15
@faho Thanks, looks like I could use command `abbr -a` to add them instead of writing the abbrs in the config.fish file, may be the documentation should advise the user avoid to add too many abbr in their config.fish file, as that would slow fish a lot (30 abbrs would cost nearly 0.6s on my machine).